Save NavigationPath to SceneStorage. (#404)
This is a little tricky. When the List selection changes, it will reset
the NavigationStack navigationPath. So when restoring state at launch in
.task, if the selectedCategory and navigationPath both change, the
navigationPath will be lost. The workaround is to hold onto the pending
navigationPath, and set it after the selectedCategory has changed. Then
forget about the pending navigationPath. Fixes #246