Separate ArchiveNavigation/NearbyModel.State json String from description and RawRepresentable (#981)
This allows the json string to be used for the CustomStringConvertible.
However can't make State RawRepresentable for a String, since State is
also Codable, and if it is RawRepresentable, there will be an infinite
loop.

This is basically getting rid of a "saving" archive log when that isn't
what was happening. Just save that for the RawRepresentable stubs, and
allow with logs to access the JSON directly.