When externally navigating to a path, set the category (#1024)
- This exposed an issue where changing both the category and path would
*clear* the path.
- In an 1-1 SwiftUI lab WWDC25, this was the known workaround suggested
by the team.
- Test URLs in a simulator with:
```
xcrun simctl openurl booted https://www.bolsinga.com/bands/ar1416.html
```
- Without this workaround, changing the category will update the UI such
that the path is cleared after the first property changes.
- When both properties change, do not animate the first category
property change to eliminate a double animation.
- When both properties change, make the path change on the next turn of
the RunLoop with an animation.
- **NOTE:** This use of DispatchQueue.main is the reason why this class
is @MainActor. If this is removed in the future, remove the annotation
from the class too.